refactor: remove performance-by-deck-helper-base.ts
